chore: double idle time but limit to 15 minutes
This commit modifies the `getEventStreamFromUrl` function in `getEventStream.ts` to double the idle time, but not more than 15 minutes. This change ensures that the idle time is increased gradually to optimize the performance of the event stream. The idle time is now calculated using the formula `idleTime *= 2` and then limited to a maximum of 15 minutes using `idleTime = Math.min(idleTime, 15 * 60 * 1000)`.

Signed-off-by: Tung Vu <me@tungv.com>