refactor: Update TypeScript compiler options
This commit updates the TypeScript compiler options in the `tsconfig.json` files for the `jerni-store-mongodb` and `jerni` packages. The changes include enabling the latest features by adding the `DOM` and `DOM.Iterable` libraries, setting the target and module to `ESNext`, and using the `react-jsx` JSX option. Additionally, some stricter flags have been enabled, such as `noUnusedLocals`, `noUnusedParameters`, and `noPropertyAccessFromIndexSignature`.

Signed-off-by: Tung Vu <me@tungv.com>